# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "DocServer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/documents/{id}/shares": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of user profile IDs that a specific document has been shared with.\n\nOnly the user who originally created (owns) the document can use this endpoint to see who they've shared it with.\nProvide the document's `id` in the URL path. Authentication via access token is required.\nIf the document hasn't been shared with anyone, it returns an empty list.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sharing"
                ],
                "summary": "See Who a Document is Shared With",
                "parameters": [
                    {
                        "type": "string",
                        "example": "doc_abc123xyz",
                        "description": "The unique identifier of the document whose share list you want to view.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the list of profile IDs the document is shared with. The 'shared_with' array contains the IDs.",
                        "schema": {
                            "$ref": "#/definitions/api.GetSharersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request: The document ID provided in the URL path is missing or invalid.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden: You are not the owner of this document, so you cannot view its share list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: No document exists with the specified ID.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while retrieving the share list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Completely replaces the list of users a specific document is shared with.\n\nProvide a JSON array named `shared_with` in the request body, containing the profile IDs of the users you want to share the document with.\n**Important:** Any users previously shared with, but *not* included in the new list, will lose access.\nTo remove *all* shares for a document, send an empty array: `{\"shared_with\": []}`.\n\nOnly the document owner can perform this operation. You cannot share a document with yourself (the owner).\nProvide the document's `id` in the URL path. Authentication via access token is required.\n\nExample Request Body (Share with user 'user_123' and 'user_456'):\n```json\n{\n\"shared_with\": [\"user_123\", \"user_456\"]\n}\n```",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sharing"
                ],
                "summary": "Set/Replace Who a Document is Shared With",
                "parameters": [
                    {
                        "type": "string",
                        "example": "doc_abc123xyz",
                        "description": "The unique identifier of the document whose share list you want to set/replace.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "A JSON object containing the 'shared_with' key, whose value is an array of profile IDs.",
                        "name": "shareRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SetSharersRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Share List Updated Successfully. No content is returned in the response body."
                    },
                    "400": {
                        "description": "Bad Request: The request body is invalid (e.g., missing 'shared_with' array, invalid JSON) OR you tried to include the owner's ID in the 'shared_with' list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden: You are not the owner of this document, so you cannot modify its share list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: No document exists with the specified ID.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while updating the share list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
        "/documents/{id}/shares/{profile_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds a single specified user (by their `profile_id`) to the list of users who can access a specific document.\n\nThis operation is *additive* – it doesn't affect other users the document might already be shared with.\nIt's also *idempotent*, meaning if you try to add a user who already has access, the operation succeeds without making any changes.\n\nOnly the document owner can perform this operation. You cannot share a document with yourself (the owner).\nProvide the document's `id` and the target user's `profile_id` in the URL path. Authentication via access token is required.",
                "tags": [
                    "Sharing"
                ],
                "summary": "Share a Document with One User",
                "parameters": [
                    {
                        "type": "string",
                        "example": "doc_abc123xyz",
                        "description": "The unique identifier of the document you want to share.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "user_123",
                        "description": "The unique identifier of the user profile you want to grant access to.",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User Added to Share List Successfully (or was already shared with). No content is returned."
                    },
                    "400": {
                        "description": "Bad Request: You tried to share the document with its owner (yourself).",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden: You are not the owner of this document, so you cannot share it.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified Document ID or Profile ID does not exist, or the IDs were missing from the URL path.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while adding the user to the share list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes a single specified user (by their `profile_id`) from the list of users who can access a specific document.\n\nThis operation only affects the specified user; other users the document is shared with remain unaffected.\nIt's *idempotent*, meaning if you try to remove a user who doesn't currently have access (or never did), the operation succeeds without error.\n\nOnly the document owner can perform this operation.\nProvide the document's `id` and the target user's `profile_id` (the one to remove) in the URL path. Authentication via access token is required.",
                "tags": [
                    "Sharing"
                ],
                "summary": "Stop Sharing a Document with One User",
                "parameters": [
                    {
                        "type": "string",
                        "example": "doc_abc123xyz",
                        "description": "The unique identifier of the document you want to modify shares for.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "user_123",
                        "description": "The unique identifier of the user profile whose access you want to revoke.",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User Removed from Share List Successfully (or was not shared with). No content is returned."
                    },
                    "401": {
                        "description": "Unauthorized: Your access token is missing, invalid, or expired.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden: You are not the owner of this document, so you cannot modify its shares.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found: The specified Document ID or Profile ID does not exist, or the IDs were missing from the URL path.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Something went wrong on the server while removing the user from the share list.",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.GetSharersResponse": {
            "type": "object",
            "properties": {
                "shared_with": {
                    "description": "List of Profile IDs (dashless)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "api.SetSharersRequest": {
            "type": "object",
            "required": [
                "shared_with"
            ],
            "properties": {
                "shared_with": {
                    "description": "Use pointer to distinguish between empty list and not provided?\nNo, binding:\"required\" means it must be present, even if empty array `[]`.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "utils.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}